+++ src/point_match.rs
@@ -1,5 +1,7 @@
 //! Point matching utilities for structured grid faces.
 
+use crate::Float;
+
 /// Find the indices of the closest point in a 2D face grid to `(x, y, z)`.
 ///
 /// The face grid is represented by flat arrays `x2`, `y2`, `z2` of length
@@ -7,21 +9,22 @@
 ///
 /// Returns `Some((u, v))` if the minimum distance is within `tol`,
 /// otherwise `None`.
+#[allow(clippy::too_many_arguments)]
 pub fn point_match(
-    x: f64,
-    y: f64,
-    z: f64,
-    x2: &[f64],
-    y2: &[f64],
-    z2: &[f64],
+    x: Float,
+    y: Float,
+    z: Float,
+    x2: &[Float],
+    y2: &[Float],
+    z2: &[Float],
     nu: usize,
     _nv: usize,
-    tol: f64,
+    tol: Float,
 ) -> Option<(usize, usize)> {
     debug_assert_eq!(x2.len(), y2.len());
     debug_assert_eq!(x2.len(), z2.len());
 
-    let mut best_dist = f64::INFINITY;
+    let mut best_dist = Float::INFINITY;
     let mut best_idx = 0usize;
 
     for (idx, ((&px, &py), &pz)) in x2.iter().zip(y2.iter()).zip(z2.iter()).enumerate() {
